WIP working through web-api tests
